i2c: cleanup and getting card UID working
This is a checkpoint where when we place the card in front of the
card reader, we have the UID printing out in the client.

However, there are multiple problems still. This code does not work
when we disable printing, in fact the printing in the driver is what
causes this to work. There are multiple hacky while loops to add a delay,
this needs to be properly fixed first.

Co-authored-by: Alwin Joshy <joshyalwin@gmail.com>