serial: cleaning up for integration with other sub-systems
This patch makes a similar change like we did for the network sub-system
by making the serial shared ring buffer it's own object file. This
makes it easier to compile and lets us just have a Makefile for
the ring buffer to compile to.

It's not clear whether this approach will be the final approach, but it
does cause less friction currently.