serial: use offsets into data regions
This patch changes the serial sub-system to always refer to
offsets into the data regions, rather than passing around
virtual addresses that are relative to each virtual address
space.

Previously, the serial sub-system relied on the data regions
being mapped into the same virtual address spaces which is
quite fragile, especially for clients that have many different
MRs mapped in compared to the driver or virtualiser.