Add explicit channel argument to timer client library
Having the timer driver's channel as a preprecessor macro means that
for systems that contain multiple programs using the timer, either
every program must have the same channel number for the timer driver,
or the client library needs to be built independently for different
programs. This may be too constricting until we develop an ergonomic
way to specify and build Microkit systems.

This commit modifies the API of the timer client library to accept the
timer's channel as a parameter.

This also has the side effect that a program could use multiple timer
devices simultaneously. Not sure how much value there is there.