Fix LWIP configuration headers
A commit was made previously that hardcoded within LWIP source code
specific lwipopts.h and arch/cc.h headers. These headers allow
users to configure LWIP. Different applications may need different
copies of them.

The proper way to handle this is to keep LWIP as it was and pick
particular copies of the headers by manipulating the include path.

This commit reverts changes to LWIP, moves the configuration headers
into a specific lwip_include directory and adds that directory to the
include path of the echo server example system.