build.zig: remove -fno-sanitize options
My fault, this should not have been merged. When we were first adding
all the build.zigs in the examples we were having runtime crashes
and I thought it was due to UBSAN so added a bunch of
`-fno-sanitize=undefined`.

I've run the serial and timer examples and they all run fine with
UBSAN enabled so I think it's fine to remove all these extra flags.
After all, we should be fixing the UB if we have any, not disabling
the UB checks.

Signed-off-by: Ivan Velickovic <i.velickovic@unsw.edu.au>