net/virtio: only set QueueNotify if we have transferred data
This is a small optimisation in the virtIO net driver that means the
virtIO net device will not be needlessly invoked if we have not been
able to enqueue any TX data. This scenario could happen under high-loads
where the virtIO net queues are full but we have more client TX data
to process.

Signed-off-by: Ivan Velickovic <i.velickovic@unsw.edu.au>