Various fixes to block subsystem
Signed-off-by: Eric Chan <eric.chan2@unsw.edu.au>

This includes fixes to some errors to the blk virt, which currently:
1. Handles the error states quite poorly, there are multiple points in the handle_client()
function that drops a request instead of enqueueing an error response.
2. Does a memcpy from client data to virtualiser-driver data. The io addresses in the
request should contain the client's data physical addresses instead of the ones in
the virtualiser-driver. The virtualiser-driver data region still has uses however for
requesting the MBR block (sector 0). The mmc and blk examples are changed to reflect this.

There are also other benign changes in various parts of the blk subsystem. These changes are
mainly to keep things consistent.
- Shuffling around the order of error checking to reduce repeated code (blk_virt).
- Improvements to comments (blk_virt).
- Removing static asserts in the blk_config files, these checks should ideally belong in the
virt and clients.
- Renaming all mentions of blk_config to blk_storage_info.
- Renaming queue_size to queue_capacity in all mentions in blk_config.
- Changing the capacity of config region (storage_info_region) to 0x1000, from 0x200000.
- Make consistent the ordering of all variable names in the sdf files for blk subsystem,
moving the "client"/"driver" portion to the last bit (e.g. blk_data_paddr_client).
- Remove REQBK_SIZE, this should just always be BLK_QUEUE_CAPACITY_DRIVER
- Rename reqbk to cached_req_info, the naming of reqbk (request book keep) was confusing imo.