blk/mmc/imx: handle virtualiser notifications while waiting for an IRQ
Previously, if the virtualiser notified the driver whilst we were
running a command (and waiting for an IRQ from the device), we acted as
if received an IRQ, not a virtualiser notification. We refactor how
notified() handles IRQs vs virtualiser notifications to ensure that
these are kept separate in future.

Signed-off-by: julia <git.ts@trainwit.ch>