build.zig: Override MICROKIT_SDK environment variable
This prevents a footgun whereby an environment variable MICROKIT_SDK is
set, but a different `-Dsdk=<sdk>` is set as an argument. This is
because the Microkit tool uses the boards path from the environment
variable preferentially to the path it is installed in.

This aligns the zig build files to the makefiles.

Signed-off-by: julia <git.ts@trainwit.ch>