Begin transition to making sDDF sub-systems agnostic to seL4 OS personality.
This commit introduces a thin-wrapper of the required seL4 APIs for sDDF
drivers and components to work. This allows sDDF to be used in other
seL4 OSes. The main other one we are targeting is the Secure Multi Server
Operating System that is a Trustworthy Systems research project. It is
based on a dynamic design and it attempting to be general purpose rather
than static like Microkit.

Currently this is done for some drivers for network, serial and timer.
Over time everything should converge to using these wrappers.

Signed-off-by: alwin-joshy <joshyalwin@gmail.com>
Signed-off-by: Ivan-Velickovic <i.velickovic@unsw.edu.au>