drivers/i2c/meson: remove README
This document is out of date and describes more than just the
driver. It describes the I2C protocol for sDDF in general as well
as specifics about the I2C device on the Odroid-C4.

For now I think it's best to remove this as the protocol is in
the design document, and I don't think it's necessary to document
details about how the hardware works (if we need to, we can do
that as code comments in the driver).

There is a question about where we write specifics about what
features the driver implements etc of the available hardware,
the i.MX8 MMC driver is an example of this. Right now there's
not a clear answer but when we begin fleshing out the developer
documentation, we can decide then.

This document is still in the commit history obviously so we
can always extract any useful bits in the future.

Signed-off-by: Ivan-Velickovic <i.velickovic@unsw.edu.au>