examples: assert sdfgen version in meta.py scripts
Until the sDDF parts of sdfgen become part of this repo, it is common
for the sdfgen version and what the metaprogram expects to become
out of sync. This is especially common now due to us frequently changing
both projects as sdfgen is still immature.

This assert means that at least the user knows if they are using the
wrong version.

Co-authored-by: Courtney Darville <courtneydarville94@outlook.com>
Signed-off-by: Ivan-Velickovic <i.velickovic@unsw.edu.au>