examples/serial: one serial for everything on QEMU virt RISC-V
Previously we, annoyingly, had to open a separate terminal
to read the virtIO console character device!

Thanks to Julia, we no longer have to do this and can instead
have on stdio that contains the input/output for both the
default (ns16650a) serial and our virtIO console serial driver.

Co-authored-by: juliab <midnight@trainwit.ch>
Signed-off-by: Ivan-Velickovic <i.velickovic@unsw.edu.au>