virtio: ack interrupts *before* handling responses
If we handle the responses before ACKing, we run the chance of having responses
appear after we (think) have processed all the responses; but then because we
have ACK'd the interrupt we will have missed them and then hang.

Linux: https://elixir.bootlin.com/linux/v6.14.6/source/drivers/virtio/virtio_mmio.c#L299-L325

This was causing hangs in the block examples and explains some of the weirdness
with our ethernet and console virtio drivers.

Signed-off-by: julia <git.ts@trainwit.ch>