drivers/timer/meson: ack IRQ during init
We observed in [1] (our regularly scheduled CI hardware tests)
that the echo server DHCP step failed on the Odroid-C2 but only
sometimes. The culprit seems to be these three lines:

2025-07-17T08:06:15.0631821Z <<seL4(CPU 0) [handleInterruptEntry/56 T0x807ffe7400 "rootserver" @8a0001a0]: Spurious interrupt!>>
2025-07-17T08:06:15.0632714Z Undelivered IRQ: 40
2025-07-17T08:06:15.0633314Z Undelivered IRQ: 238

where 238 is the timer IRQ. This caused the timer driver to not
operate properly as the IRQ was delivered before it started and
therefore was masked by the kernel.

I believe the undelivered IRQs themselves are not a concern as they
could genuinely happen depending on what the previous bootloader was
doing, in this case U-Boot.

In general our drivers should be more robust to this kind of thing
but this commit is specifically for the Odroid's timer driver.

A future commit will fixup all the drivers.

[1]: https://trustworthy.systems/github/ci-artifacts/sddf/2025-07/Run2079_f772ebd_Examples/logs/4_Run%20%28hardware%29.txt.html

Signed-off-by: Ivan-Velickovic <i.velickovic@unsw.edu.au>