drivers/network/dwmac-5.10a: use BIT macro
I ran into the UBSAN checks going off when I was trying to get
virtIO network working with the HiFive P550 which uses this as the
native driver.

The problem is certain defines such as `DESC_RXCTRL_RXINTDIS` have
`1 << 31` which means it's shifting a 32-bit *signed* integer which
by that many bits which is undefined behaviour.

We already use the BIT macro in various places in that header so
just change it to use it everywhere.

Signed-off-by: Ivan Velickovic <i.velickovic@unsw.edu.au>