hack: skip input events with type 0
I don't know why but for every IRQ we get it seems that
event_virtq.used->idx increases by 2 for keyboard, one
descriptor contains the actual keyboard press and then the
next is just all zeroes for some reason.

Need to look at QEMU source code to understand if it's
intentional for the device or a bug in the driver.

Signed-off-by: Ivan Velickovic <i.velickovic@unsw.edu.au>