ci: use python3.13 explicitly on github hosted macos
What a fantastic situation! You can pin the version of macOS
to use for CI where the default version of Python is perfectly fine
to change at any given moment.

Python 3.14 only got recently I believe last week or the week before
so sdfgen does not have a package for it yet. I will fix that soon.

There also seems to be some issues with 3.14 and our CI scripts.

Signed-off-by: Ivan Velickovic <i.velickovic@unsw.edu.au>