benchmark: get cycle counts and utilisation info on RISC-V working
Right now all of our benchmarking has been done on ARM, and with
the HiFive P550 we have a more mature RISC-V platform that we can
test the DWMAC ethernet driver with.

sel4bench.h is very ARM oriented but that header is mainly concerned
with getting PMU event information that is not done by the kernel
(cache misses etc).

Right now this patch refactors benchmark.c to be a bit more legible
and allows CPU utilisation to be recorded on RISC-V platforms.

I have in-progress changes to do proper PMU access on RISC-V for
the other events but I want to get this patch merged first as
it allows Kurt to do a performance analysis on RISC-V for his changes
to the DWMAC ethernet driver.

Signed-off-by: Ivan Velickovic <i.velickovic@unsw.edu.au>