ci/matrix.py: cheshire -> serengeti
Since [1] in seL4/machine_queue, we cannot use cheshire1/cheshire2 which
is causing CI to fail right now because we're trying to run on non-existent
platforms right now.

For our sDDF CI purposes we want to use Serengeti anyways since it has
more peripherals and what we are using internally, so remove all cheshire
run-time tests.

[1]: https://github.com/seL4/machine_queue/commit/f0835aa54749667d30c9143e0fd3353e9367f2dd

Signed-off-by: Ivan Velickovic <i.velickovic@unsw.edu.au>