Revert commented lines in libc
0911c90aa7c39a26de52f566b422cc530605bd4e commented these lines
and I don't understand why, this is vendored code so we should
just leave it as is. I believe it was commented out due to a
misunderstanding that these were used instead of the optimised
assembly version. This is only true on RISC-V where an optimised
assembly version does not (yet) exist in newlibc.

Signed-off-by: Ivan Velickovic <i.velickovic@unsw.edu.au>